Add storybook and some different stories
