+++ package.json
@@ -25,6 +25,12 @@
     "@babel/polyfill": "^7.10.4",
     "@babel/preset-env": "^7.10.4",
     "@babel/preset-react": "^7.10.4",
+    "@storybook/addon-actions": "^6.4.8",
+    "@storybook/addon-essentials": "^6.4.8",
+    "@storybook/addon-links": "^6.4.8",
+    "@storybook/builder-webpack5": "^6.4.8",
+    "@storybook/manager-webpack5": "^6.4.8",
+    "@storybook/react": "^6.4.8",
     "@types/d3-ease": "^3.0.0",
     "@types/exenv": "^1.2.0",
     "@types/react": "^16.9.56",
@@ -80,11 +86,13 @@
     "start": "webpack-dev-server --mode development",
     "test": "jest test --config jest.unit.config.js",
     "test-e2e": "jest test --config jest.e2e.config.js",
-    "version": "npm run build"
+    "version": "npm run build",
+    "storybook": "start-storybook -p 6006",
+    "build-storybook": "build-storybook"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/kenwheeler/nuka-carousel.git"
+    "url": "https://github.com/FormidableLabs/nuka-carousel.git"
   },
   "keywords": [
     "react",
